Refactor data products into product modules
