+++ data/fonts/core/src/main/kotlin/ee/schimke/composeai/data/fonts/FontsUsedDataProduct.kt
@@ -1,0 +1,42 @@
+package ee.schimke.composeai.data.fonts
+
+import java.io.File
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.Json
+
+/** Path-backed producer for `fonts/used`, written by backend render loops in default mode. */
+object FontsUsedDataProducer {
+  const val KIND: String = "fonts/used"
+  const val SCHEMA_VERSION: Int = 1
+  const val FILE: String = "fonts-used.json"
+
+  val json: Json = Json {
+    encodeDefaults = true
+    ignoreUnknownKeys = true
+    prettyPrint = false
+  }
+
+  fun readPayload(rootDir: File, previewId: String): FontsUsedPayload? {
+    val file = rootDir.resolve(previewId).resolve(FILE)
+    if (!file.exists()) return null
+    return json.decodeFromString(FontsUsedPayload.serializer(), file.readText())
+  }
+
+  fun writeArtifacts(rootDir: File, previewId: String, payload: FontsUsedPayload) {
+    val previewDir = rootDir.resolve(previewId).also { it.mkdirs() }
+    previewDir.resolve(FILE).writeText(json.encodeToString(FontsUsedPayload.serializer(), payload))
+  }
+}
+
+@Serializable data class FontsUsedPayload(val fonts: List<FontUsedEntry>)
+
+@Serializable
+data class FontUsedEntry(
+  val requestedFamily: String,
+  val resolvedFamily: String,
+  val weight: Int,
+  val style: String,
+  val sourceFile: String? = null,
+  val fellBackFrom: List<String>? = null,
+  val consumerNodeIds: List<String> = emptyList(),
+)
