+++ data/layoutinspector/core/src/main/kotlin/ee/schimke/composeai/data/layoutinspector/ComposeSemanticsModels.kt
@@ -1,0 +1,30 @@
+package ee.schimke.composeai.data.layoutinspector
+
+import kotlinx.serialization.Serializable
+
+object ComposeSemanticsProduct {
+  const val KIND: String = "compose/semantics"
+  const val SCHEMA_VERSION: Int = 1
+  const val FILE: String = "compose-semantics.json"
+}
+
+@Serializable data class ComposeSemanticsPayload(val root: ComposeSemanticsNode)
+
+@Serializable
+data class ComposeSemanticsNode(
+  val nodeId: String,
+  val boundsInRoot: String,
+  val label: String? = null,
+  val text: String? = null,
+  val layoutText: String? = null,
+  val layoutFontSize: String? = null,
+  val layoutForegroundColor: String? = null,
+  val layoutBackgroundColor: String? = null,
+  val editableText: String? = null,
+  val inputText: String? = null,
+  val role: String? = null,
+  val testTag: String? = null,
+  val mergeMode: String? = null,
+  val clickable: Boolean = false,
+  val children: List<ComposeSemanticsNode> = emptyList(),
+)
