+++ data/recomposition/connector/build.gradle.kts
@@ -1,0 +1,21 @@
+plugins {
+  alias(libs.plugins.kotlin.jvm)
+  alias(libs.plugins.compose.multiplatform)
+  alias(libs.plugins.compose.compiler)
+  alias(libs.plugins.kotlin.serialization)
+}
+
+group = "ee.schimke.composeai"
+
+version = providers.environmentVariable("PLUGIN_VERSION").orNull ?: "0.0.0-SNAPSHOT"
+
+dependencies {
+  api(project(":daemon:core"))
+  implementation(compose.runtime)
+  implementation(compose.ui)
+  testImplementation(libs.junit)
+}
+
+java { toolchain { languageVersion.set(JavaLanguageVersion.of(17)) } }
+
+tasks.withType<Test>().configureEach { useJUnit() }
